Extract contact form to its own component
